+++ app/src/main/java/com/prog/tlc/btexchange/protocollo/RouteRequest.java
@@ -7,15 +7,15 @@ import java.io.Serializable;
  */
 public class RouteRequest implements Serializable{
     private String source_addr;
-    private int source_sequence_number; //identifica univocamente questo messaggio RREQ
+    private long source_sequence_number; //identifica univocamente questo messaggio RREQ
     private String dest_addr;
-    private int dest_sequence_number; /*The latest sequence number received in the past
+    private long dest_sequence_number; /*The latest sequence number received in the past
                                         by the originator for any route towards the
                                         destination.*/
     private int hop_cnt;
     private String last_sender; /*indirizzo MAC del nodo che ha appena inviato questa RREQ*/
 
-    public RouteRequest(String source_addr, int source_sequence_number, String dest_addr, int dest_sequence_number, int hop_cnt, String last_sender) {
+    public RouteRequest(String source_addr, long source_sequence_number, String dest_addr, long dest_sequence_number, int hop_cnt, String last_sender) {
         this.source_addr = source_addr;
         this.source_sequence_number = source_sequence_number;
         this.dest_addr = dest_addr;
@@ -29,7 +29,7 @@ public class RouteRequest implements Serializable{
         return source_addr;
     }
 
-    public int getSource_sequence_number() {
+    public long getSource_sequence_number() {
         return source_sequence_number;
     }
 
@@ -37,7 +37,7 @@ public class RouteRequest implements Serializable{
         return dest_addr;
     }
 
-    public int getDest_sequence_number() {
+    public long getDest_sequence_number() {
         return dest_sequence_number;
     }
